+++ tools/analyzer/lsp.h
@@ -918,6 +918,63 @@ struct DocumentLink {
   }
 };
 
+struct DocumentSymbolParams : public WorkDoneProgressParams, public PartialResultParams {
+  TextDocumentIdentifier textDocument;
+
+  template <typename T>
+  void jsonify(T &t) {
+    WorkDoneProgressParams::jsonify(t);
+    PartialResultParams::jsonify(t);
+    JSONIFY(textDocument);
+  }
+};
+
+enum class SymbolKind : unsigned int {
+  File = 1,
+  Module = 2,
+  Namespace = 3,
+  Package = 4,
+  Class = 5,
+  Method = 6,
+  Property = 7,
+  Field = 8,
+  Constructor = 9,
+  Enum = 10,
+  Interface = 11,
+  Function = 12,
+  Variable = 13,
+  Constant = 14,
+  String = 15,
+  Number = 16,
+  Boolean = 17,
+  Array = 18,
+  Object = 19,
+  Key = 20,
+  Null = 21,
+  EnumMember = 22,
+  Struct = 23,
+  Event = 24,
+  Operator = 25,
+  TypeParameter = 26,
+};
+
+struct DocumentSymbol {
+  std::string name;
+  Optional<std::string> detail;
+  SymbolKind kind;
+  Range range;
+  Range selectionRange;
+
+  template <typename T>
+  void jsonify(T &t) {
+    JSONIFY(name);
+    JSONIFY(detail);
+    JSONIFY(kind);
+    JSONIFY(range);
+    JSONIFY(selectionRange);
+  }
+};
+
 #undef JSONIFY
 
 } // namespace ydsh::lsp